chore: codebase cleanup — dead code removal, decimal precision fix, error visibility
- Removed 9 dead frontend files + 14 orphaned CSS classes + 1 dead backend schema + dead
  core/tasks.py (unreferenced anywhere, "not awaitable" Pyrefly warnings were false positives
  from redis-py's imprecise async/sync stub typing — same pattern used unflagged in
  auth_service.py/rate_limiter.py)
- Fixed real bug: member.balance was float, risking accumulated floating-point drift across
  successive adjustments — migrated to Decimal, added precision regression test
- Normalized cash_register_service exception handling: close() now raises AppException(409)
  matching add_withdrawal() and the rest of the codebase's precondition-failure convention,
  instead of the outlier NotFoundException(404)
- Added missing onError handlers to withdrawMutation/closeMutation in ReceptionPage — errors
  were previously silent to the user
- Installed and configured ESLint (react-hooks/rules-of-hooks + exhaustive-deps) — zero hooks
  findings, confirms prior manual audit was accurate
- Wired CheckInCard.onGoToPayments for frozen/expired members; added a new debt-block-specific
  CTA (distinguishes the debt 409 from the already-checked-in-today 409) that deep-links to
  Payments with the member preselected via ?member= query param
- Fixed a real bug caught during this work: two onClick={openCreate} call sites would have
  leaked the MouseEvent as the member argument after openCreate's signature changed
- 154/154 backend tests passing, ruff clean, tsc clean, eslint clean (2 pre-existing incidental
  findings unrelated to hooks, not touched)
- Verified live in browser: debt-block and frozen/expired paths both tested end-to-end with
  real data, confirmed no side effects (no check-in created when blocked)

+++ backend/app/models/member.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import date
+from decimal import Decimal
 
 from sqlalchemy import Date, ForeignKey, Numeric, String, Text
 from sqlalchemy.dialects.postgresql import UUID
@@ -23,7 +24,7 @@ class Member(Base, TimestampMixin, SoftDeleteMixin, UUIDMixin):
     notes: Mapped[str | None] = mapped_column(Text, nullable=True)
     status: Mapped[str] = mapped_column(String(20), default="active")
     # Cached sum of MemberBalanceMovement.amount. Negative = deudor, positive = a favor.
-    balance: Mapped[float] = mapped_column(Numeric(10, 2), default=0, nullable=False)
+    balance: Mapped[Decimal] = mapped_column(Numeric(10, 2), default=0, nullable=False)
 
     gym = relationship("Gym", back_populates="members")
     memberships = relationship("MemberMembership", back_populates="member", lazy="selectin")
